drop phpdbg_opcode.{c,h}
