+++ sapi/phpdbg/phpdbg_prompt.c
@@ -32,7 +32,6 @@
 #include "phpdbg_print.h"
 #include "phpdbg_info.h"
 #include "phpdbg_break.h"
-#include "phpdbg_opcode.h"
 #include "phpdbg_list.h"
 #include "phpdbg_utils.h"
 #include "phpdbg_prompt.h"
